week1/프로필만들기/requirements.txt
authlib>=1.3.1
fastapi>=0.115.0
jinja2>=3.1.6
numpy>=1.26.0
Pillow>=10.4.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
uvicorn>=0.34.0
